Refactor shared to plugins; Moved helpers to utils
