+++ packages/plugins/src/fix-transform/index.ts
@@ -1,22 +1,13 @@
 import ts from 'typescript';
 import { type RehearsalService } from '@rehearsal/service';
-import { getCommentsOnlyResult } from '../helpers/transform-utils';
+import { getCodemodResult, getCommentsOnlyResult } from './utils';
+import { type FixResult, type FixedFile, FixTransform as FixTransformI } from './interfaces';
 
-export interface FixedFile {
-  fileName: string;
-  updatedText?: string;
-  location: {
-    line: number;
-    character: number;
-  };
-}
 
-export interface FixResult {
-  fixedFiles: FixedFile[];
-  commentedFiles: FixedFile[];
-}
+export { getCodemodResult, getCommentsOnlyResult };
+export { FixResult, FixedFile };
 
-export class FixTransform {
+export class FixTransform implements FixTransformI {
   /** Engineer friendly message describes steps needs to be done to fix the related issue */
   hint?: string;
 
@@ -27,3 +18,4 @@ export class FixTransform {
     return this.fix ? this.fix(diagnostic, service) : getCommentsOnlyResult(diagnostic);
   }
 }
+
