--- packages/shared/src/index.ts
@@ -1,3 +1,0 @@
-export * from './interfaces';
-export * from './helpers';
-export * from './plugins';
