--- packages/shared/src/interfaces/index.ts
@@ -1,1 +1,0 @@
-export * from './fix-transform';
