+++ packages/upgrade/package.json
@@ -31,7 +31,7 @@
   "dependencies": {
     "@rehearsal/reporter": "^0.0.27",
     "@rehearsal/service": "^0.0.27",
-    "@rehearsal/shared": "^0.0.27",
+    "@rehearsal/plugins": "^0.0.27",
     "@rehearsal/utils": "^0.0.27",
     "commander": "^9.4.0",
     "eslint": "^8.11.0",
