+++ packages/upgrade/src/plugins/index.ts
@@ -1,2 +1,2 @@
-export { EmptyLinesPreservePlugin, EmptyLinesRestorePlugin, LintPlugin } from '@rehearsal/shared';
+export { EmptyLinesPreservePlugin, EmptyLinesRestorePlugin, LintPlugin } from '@rehearsal/plugins';
 export { DiagnosticAutofixPlugin } from './diagnostics-autofix.plugin';
