+++ packages/upgrade/src/transforms/2322-fix-transform.ts
@@ -1,8 +1,8 @@
 import ts from 'typescript';
 
-import { FixTransform, type FixResult } from '@rehearsal/shared';
-import { getCommentsOnlyResult } from '@rehearsal/shared';
-import { transformDiagnosedNode } from '@rehearsal/shared';
+import { FixTransform, type FixResult } from '@rehearsal/plugins';
+import { getCommentsOnlyResult } from '@rehearsal/plugins';
+import { transformDiagnosedNode } from '@rehearsal/utils';
 
 export class FixTransform2322 extends FixTransform {
   hint = `Type '{0}' is being returned or assigned, but type '{1}' is expected. Please convert type '{0}' to type '{1}', or return or assign a variable of type '{1}'`;
