+++ packages/upgrade/src/transforms/2790-fix-transform.ts
@@ -2,9 +2,9 @@ import ts from 'typescript';
 
 import { type RehearsalService } from '@rehearsal/service';
 
-import { FixTransform, type FixResult } from '@rehearsal/shared';
-import { getCommentsOnlyResult, getCodemodResult } from '@rehearsal/shared';
-import { findNodeAtPosition, insertIntoText } from '@rehearsal/shared';
+import { FixTransform, type FixResult } from '@rehearsal/plugins';
+import { getCommentsOnlyResult, getCodemodResult } from '@rehearsal/plugins';
+import { findNodeAtPosition, insertIntoText } from '@rehearsal/utils';
 
 import {
   getClassByName,
