+++ packages/utils/src/helpers/index.ts
@@ -1,3 +1,2 @@
 export * from './strings';
-export * from './transform-utils';
 export * from './typescript-ast';
