+++ packages/utils/src/helpers/strings.ts
